abstract out _merge
